fix: Main project의 integration test packae명 오류 수정 - Global이 아닌 package들이 global 하위에 속했던 버그 픽스
