+++ src/integration/java/com/github/can019/base/test/util/listener/performance/ParallelTestTimeExecutionListener.java
@@ -1,6 +1,6 @@
-package com.github.can019.base.global.test.util.listener.performance;
+package com.github.can019.base.test.util.listener.performance;
 
-import com.github.can019.base.global.test.util.report.StopWatchReporter;
+import com.github.can019.base.test.util.report.StopWatchReporter;
 import org.springframework.test.context.TestContext;
 import org.springframework.test.context.support.AbstractTestExecutionListener;
 import org.springframework.util.StopWatch;
@@ -8,7 +8,7 @@ import org.springframework.util.StopWatch;
 import java.text.SimpleDateFormat;
 import java.util.Date;
 
-import static com.github.can019.base.global.test.util.report.StopWatchReporter.*;
+import static com.github.can019.base.test.util.report.StopWatchReporter.*;
 
 public class ParallelTestTimeExecutionListener extends AbstractTestExecutionListener {
     private static final ThreadLocal<StopWatch> threadLocalStopWatch =
