+++ src/integration/java/com/github/can019/base/test/util/url/BaseUrl.java
@@ -1,4 +1,4 @@
-package com.github.can019.base.global.test.util.url;
+package com.github.can019.base.test.util.url;
 
 public enum BaseUrl {
     LOCALHOST_URL("http://localhost");
